Split out game status checks and add piece colour search
