+++ test/chess/board_test.exs
@@ -22,6 +22,17 @@ defmodule Chess.BoardTest do
     assert Board.search(board, piece) == expected_result
   end
 
+  test "finds pieces on the board with a partial search" do
+    board = Board.default
+
+    piece = %{"colour" => "white"}
+    expected_result = [
+      {0, 1}, {1, 1}, {2, 1}, {3, 1}, {4, 1}, {5, 1}, {6, 1}, {7, 1},
+      {0, 0}, {1, 0}, {2, 0}, {3, 0}, {4, 0}, {5, 0}, {6, 0}, {7, 0},
+    ] |> Enum.sort
+    assert Board.search(board, piece) == expected_result
+  end
+
   test "finds a single piece on the board" do
     board = Board.default
 
@@ -39,41 +50,5 @@ defmodule Chess.BoardTest do
     assert new_board == %{
       "5,2" => %{"type" => "queen", "colour" => "white"},
     }
-  end
-
-  test "recognise when the king is in check" do
-    board = %{
-      "4,0" => %{"type" => "king", "colour" => "white"},
-      "4,7" => %{"type" => "queen", "colour" => "black"},
-    }
-
-    assert Board.king_in_check?(board, "white")
-  end
-
-  test "recognise when the king is not in check" do
-    board = %{
-      "5,0" => %{"type" => "king", "colour" => "white"},
-      "4,7" => %{"type" => "queen", "colour" => "black"},
-    }
-
-    refute Board.king_in_check?(board, "white")
-  end
-
-  test "recognize when the king is in check by a knight" do
-    board = %{
-      "4,0" => %{"type" => "king", "colour" => "white"},
-      "3,2" => %{"type" => "knight", "colour" => "black"},
-    }
-
-    assert Board.king_in_check?(board, "white")
-  end
-
-  test "recognize when the king is in check by a pawn" do
-    board = %{
-      "4,0" => %{"type" => "king", "colour" => "white"},
-      "3,1" => %{"type" => "pawn", "colour" => "black"},
-    }
-
-    assert Board.king_in_check?(board, "white")
   end
 end
